refactor: Chats 구조 분리
